+++ xmlloader/src/test/java/org/killbill/xmlloader/TestXMLWriter.java
@@ -41,8 +41,9 @@ public class TestXMLWriter {
         final InputStream is = new ByteArrayInputStream(TEST_XML.getBytes());
         final XmlTestClass test = XMLLoader.getObjectFromStream(is, XmlTestClass.class);
         assertEquals(test.getFoo(), "foo");
-        assertEquals(test.getBar(), 1.0);
-        assertEquals(test.getLala(), 42);
+        //noinspection RedundantCast
+        assertEquals((double) test.getBar(), 1.0);
+        assertEquals((double) test.getLala(), 42);
 
         final String output = XMLWriter.writeXML(test, XmlTestClass.class);
         //System.out.println(output);
